#17 カスタムフックの導入
